+++ app/src/main/java/com/example/chillinapp/data/stress/StressData.kt
@@ -2,7 +2,7 @@ package com.example.chillinapp.data.stress
 
 data class StressRawData(
     val timestamp: Long = 0,
-    val heartrateSensor: Double = 0.0,
+    val heartrateSensor: Float = 0.0f,
     val skinTemperatureSensor: Double = 0.0
 )
